New endpoint PUT report that regenerates report html from data BD
