+++ core/src/main/java/com/github/jorge2m/testmaker/testreports/stepstore/StepEvidence.java
@@ -3,6 +3,7 @@ package com.github.jorge2m.testmaker.testreports.stepstore;
 import java.io.File;
 
 import com.github.jorge2m.testmaker.domain.suitetree.StepTM;
+import com.github.jorge2m.testmaker.domain.suitetree.TestCaseBean;
 
 public enum StepEvidence {
 	IMAGEN("png", "icon-hardcopy.png", "Page Hardcopy"), 
@@ -24,6 +25,17 @@ public enum StepEvidence {
 	public String getPathFile(StepTM step) {
 		String fileName = getNameFileEvidence(step);
 		return (step.getPathDirectory() + File.separator + fileName);
+	}
+	
+	public String getPathFile(TestCaseBean testCase, StepTM step) {
+		String fileName = getNameFileEvidence(step);
+		return (testCase.getTestPathDirectory() + File.separator + fileName);
+	}
+	
+	public boolean fileExists(TestCaseBean testCase, StepTM step) {
+		String pathFileStep = getPathFile(testCase, step);
+		File indexFile = new File(pathFileStep);
+		return indexFile.exists();
 	}
 	
 	public boolean fileExists(StepTM step) {
